pyproject.toml
[project]
name = "RAG-QA"
version = "1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu==1.7.4",
    "httpx==0.27.0",
    "numpy==1.26.4",
    "openai==1.12.0",
    "pdfplumber==0.10.3",
    "pillow==10.2.0",
    "pydantic==2.6.1",
    "pymupdf==1.23.26",
    "pytesseract==0.3.10",
    "pytest==8.1.1",
    "python-dotenv==1.0.1",
    "regex==2023.12.25",
    "reportlab==4.1.0",
    "requests==2.31.0",
    "torch==2.2.1",
    "torchaudio==2.2.1",
    "torchvision==0.17.1",
    "tqdm==4.66.1",
    "transformers==4.38.2",
    "typing-extensions>=4.8.0",
]
